dailybuild: trivial warnings の抑止（libraries, src/tests)
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@6773 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Foundation/UTClapack.h
@@ -150,7 +150,7 @@ template <class AD, class BD>
 int sprsygvx(const PTM::MatrixImp<AD>& a, const PTM::MatrixImp<AD>& b, PTM::VectorImp<BD>& e, PTM::MatrixImp<AD>& v, int start, int interval){
 	e.clear();
 	v.clear();
-	int size = a.height();
+	int size = (int)a.height();
 	assert(a.width() == size);
 	assert(b.height() == size || b.width() == size);
 	assert(e.size() == interval);
