+++ src/Physics/PHContactDetector.h
@@ -511,7 +511,7 @@ public:
 				}
 				cur.insert(it->index);
 #ifdef _DEBUG
-				if (nMaxOverlapObject < (int)cur.size()) nMaxOverlapObject = cur.size();
+				if (nMaxOverlapObject < (int)cur.size()) nMaxOverlapObject = (int)cur.size();
 #endif
 			}else{
 				cur.erase(it->index);			//	終端なので削除．
@@ -589,7 +589,7 @@ public:
 				// 初めはitfが無いからここから始まる
 				cur.insert(it->index);
 #ifdef _DEBUG
-				if (nMaxOverlapObject < (int)cur.size()) nMaxOverlapObject = cur.size();
+				if (nMaxOverlapObject < (int)cur.size()) nMaxOverlapObject = (int)cur.size();
 #endif
 			}else{
 				cur.erase(it->index);			//	終端なので削除．
